+++ amqp_job_queue.go
@@ -104,10 +104,10 @@ func (q *AMQPJobQueue) Jobs(ctx gocontext.Context) (outChan <-chan Job, err erro
 
 				err := json.Unmarshal(delivery.Body, buildJob.payload)
 				if err != nil {
-					logger.WithField("err", err).Error("payload JSON parse error, attempting to nack delivery")
+					logger.WithField("err", err).Error("payload JSON parse error, attempting to ack+drop delivery")
 					err := delivery.Ack(false)
 					if err != nil {
-						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't nack delivery")
+						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't ack+drop delivery")
 					}
 					continue
 				}
@@ -116,20 +116,20 @@ func (q *AMQPJobQueue) Jobs(ctx gocontext.Context) (outChan <-chan Job, err erro
 
 				err = json.Unmarshal(delivery.Body, &startAttrs)
 				if err != nil {
-					logger.WithField("err", err).Error("start attributes JSON parse error, attempting to nack delivery")
+					logger.WithField("err", err).Error("start attributes JSON parse error, attempting to ack+drop delivery")
 					err := delivery.Ack(false)
 					if err != nil {
-						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't nack delivery")
+						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't ack+drop delivery")
 					}
 					continue
 				}
 
 				buildJob.rawPayload, err = simplejson.NewJson(delivery.Body)
 				if err != nil {
-					logger.WithField("err", err).Error("raw payload JSON parse error, attempting to nack delivery")
+					logger.WithField("err", err).Error("raw payload JSON parse error, attempting to ack+drop delivery")
 					err := delivery.Ack(false)
 					if err != nil {
-						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't nack delivery")
+						logger.WithField("err", err).WithField("delivery", delivery).Error("couldn't ack+drop delivery")
 					}
 					continue
 				}